requirements.txt
faiss-cpu
sentence-transformers
pandas
numpy
requests
